Split unary and streaming into separate files, add initial streaming code
