+++ Cargo.toml
@@ -21,7 +21,7 @@ futures = "0.3.30"
 ginepro = "0.7.2"
 hyper = { version = "1.0.1", features = ["http1", "http2", "server"] }
 hyper-util = { version = "0.1.5", features = ["server-auto", "server-graceful", "tokio"] }
-mio = "0.8.11"
+mio = "1.0.0"
 prost = "0.12.3"
 reqwest = { version = "0.12.4", features = ["blocking", "rustls-tls", "json"] }
 rustls = {version = "0.23.9", default-features = false, features = ["std"]}
@@ -29,17 +29,18 @@ rustls-pemfile = "2.1.0"
 rustls-webpki = "0.102.2"
 serde = { version = "1.0.200", features = ["derive"] }
 serde_json = "1.0.116"
-serde_yml = "0.0.5"
+serde_yml = "0.0.10"
 thiserror = "1.0.60"
 tokio = { version = "1.37.0", features = ["rt", "rt-multi-thread", "parking_lot", "signal", "sync", "fs"] }
 tokio-rustls = { version = "0.26.0" }
-tokio-stream = "0.1.14"
+tokio-stream = { version = "0.1.15", features = ["sync"] }
 tonic = { version = "0.11.0", features = ["tls"] }
 tower-service = "0.3"
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["json", "env-filter"] }
 url = "2.5.0"
 uuid = { version = "1.8.0", features = ["v4", "fast-rng"] }
+async-trait = "0.1.80"
 
 [build-dependencies]
 tonic-build = "0.11.0"
