+++ src/server.rs
@@ -15,9 +15,7 @@ use axum::{
 use futures::StreamExt;
 use hyper::body::Incoming;
 use hyper_util::rt::{TokioExecutor, TokioIo};
-use rustls::server::WebPkiClientVerifier;
-use rustls::RootCertStore;
-use rustls::ServerConfig;
+use rustls::{server::WebPkiClientVerifier, RootCertStore, ServerConfig};
 use tokio::{net::TcpListener, signal};
 use tokio_rustls::TlsAcceptor;
 use tower_service::Service;
